perf : redis use pipeline and optimise logic

+++ cmd/interaction/service/delete_comment.go
@@ -4,15 +4,20 @@ import (
 	"github.com/ozline/tiktok/cmd/interaction/dal/cache"
 	"github.com/ozline/tiktok/cmd/interaction/dal/db"
 	"github.com/ozline/tiktok/kitex_gen/interaction"
+	"github.com/ozline/tiktok/pkg/errno"
 	"strconv"
 )
 
 // DeleteComment delete comment
-func (s *InteractionService) DeleteComment(req *interaction.CommentActionRequest) (*db.Comment, error) {
+func (s *InteractionService) DeleteComment(req *interaction.CommentActionRequest, userId int64) (*db.Comment, error) {
 
 	comment, err := db.GetCommentByID(s.ctx, *req.CommentId)
 	if err != nil {
 		return nil, err
+	}
+
+	if comment.UserId != userId {
+		return nil, errno.AuthorizationFailedError
 	}
 
 	comment, err = db.DeleteComment(s.ctx, comment)
